% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bulast.R
\name{bulast}
\alias{bulast}
\title{Request data from the Bureau of Labor Statistics API}
\usage{
bulast(seriesid, startyear = NULL, endyear = NULL, registrationKey = NULL,
  catalog = NULL, calculations = NULL, annualaverage = NULL)
}
\arguments{
\item{seriesid}{character vector of series identifiers}

\item{startyear}{integer start year for the data}

\item{endyear}{integer end year for the data (maximum of 20 years returned)}

\item{registrationKey}{BLS API registration key, required for V2 API. Freely
available from here: http://data.bls.gov/registrationEngine/}

\item{catalog}{boolean include catalog data (currently not returned); requires
registrationKey}

\item{calculations}{boolean include calculations data; requires registrationKey}

\item{annualaverage}{boolean include annual average ("M13") data; requires
  registrationKey}
}
\value{
list with names status (character), responseTime (integer), message
  (list of character), Results (data.table)
}
\description{
Request data from the Bureau of Labor Statistics API
}
\examples{
r <- bulast(c('LAUCN040010000000005', 'LAUCN040010000000006'))
r$Results
}

